requirements.txt
# Core Data Libraries
numpy==2.1.3
pandas>=2.2.0
python-dotenv==1.0.1
filetype==1.2.0

# API Wrappers (Updated for 2025)
openai>=1.58.1
tweepy>=4.16.0
praw>=7.8.1
yfinance>=0.2.50  # Updated from 1.0 for better stability
requests>=2.32.3
beautifulsoup4>=4.12.3
flask>=3.0.0      # Added for microservice functionality